Update maximal-computing.md
Felt like there were a lot of fragments and non sequitors so I've tried to weave it together more, and also create recommendations and suggestions. Trying to bring out proportionality as a thread too. I still think it could do with a bit more in terms of referencing relevant literature

I still like CALLING it "maximal computing" as a section, but I wonder if we really need to use the term throughout? Maybe we should just keep the title but go with HPC in this section?